+++ wand/api.py
@@ -97,6 +97,12 @@ library.MagickToMime.restype = ctypes.POINTER(ctypes.c_char)
 library.MagickGetImageSignature.argtypes = [ctypes.c_void_p]
 library.MagickGetImageSignature.restype = ctypes.c_char_p
 
+library.MagickGetImageBackgroundColor.argtypes = [ctypes.c_void_p,
+                                                  ctypes.c_void_p]
+
+library.MagickSetImageBackgroundColor.argtypes = [ctypes.c_void_p,
+                                                  ctypes.c_void_p]
+
 library.MagickGetImageBlob.argtypes = [ctypes.c_void_p,
                                        ctypes.POINTER(ctypes.c_size_t)]
 library.MagickGetImageBlob.restype = ctypes.POINTER(ctypes.c_ubyte)