+++ docs/changes.rst
@@ -6,6 +6,8 @@ Version 0.1.9
 
 To be released. Still alpha version.
 
+- Added :attr:`Image.background_color <wand.image.Image.background_color>`
+  property.
 - Added ``==`` operator for :class:`~wand.image.Image` type.
 - Added :func:`hash()` support of :class:`~wand.image.Image` type.
 - Added :attr:`Image.signature <wand.image.Image.signature>` property.